requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.3
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djoser==2.0.5
gitdb==4.0.9
GitPython==3.1.26
inflection==0.3.1
iniconfig==1.1.1
lazy-object-proxy==1.7.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytest-assert-utils==0.3.0
pytest-common-subject==1.0.5
pytest-drf==1.1.2
pytest-fixture-order==0.1.3
pytest-lambda==1.2.5
pytz==2021.3
smmap==5.0.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
wrapt==1.13.3
